+++ Questionable/Controller/Steps/Interactions/Duty.cs
@@ -29,15 +29,18 @@ internal static class Duty
             bool allowUnsync = step.DutyOptions.CanUnsync is not false;
             AutoDutyIpc.DutyMode dutyMode = AutoDutyIpc.DutyMode.Support;
             if (allowUnsync && quest.Id is QuestId { Value: >= 357 and <= 360 })
-                dutyMode = AutoDutyIpc.DutyMode.UnsyncRegular;
+                dutyMode = AutoDutyIpc.DutyMode.Regular;
+            else if (quest.Id is QuestId { Value: 4646 or 4733 or 4789 or 5441})
+                dutyMode = AutoDutyIpc.DutyMode.Variant;
             else if (allowUnsync && configuration.Duties.RunUnsynced)
             {
                 unsafe
                 {
                     if (territoryData.TryGetContentFinderCondition(step.DutyOptions.ContentFinderConditionId,
                                                                    out TerritoryData.ContentFinderConditionData? cfcData) &&
-                            PlayerState.Instance()->CurrentLevel - 20 >= cfcData.ClassJobLevelSync)
-                        dutyMode = AutoDutyIpc.DutyMode.UnsyncRegular;
+                            PlayerState.Instance()->CurrentLevel - 20 >= cfcData.ClassJobLevelSync &&
+                            !cfcData.ContentType.Equals(EContentType.Trials))
+                        dutyMode = AutoDutyIpc.DutyMode.Regular;
                 }
             }
 
@@ -124,7 +127,7 @@ internal static class Duty
                     currentItemLevel - 200 >= cfcData.RequiredItemLevel &&
                     !cfcData.ContentType.Equals(EContentType.Trials))
                 {
-                    dutyMode = AutoDutyIpc.DutyMode.UnsyncRegular;
+                    dutyMode = AutoDutyIpc.DutyMode.Regular;
                 }
             }